# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Virello Studio API",
    "description": "Turn long videos into short, editable clips.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_id}/clips": {
      "get": {
        "tags": [
          "clips"
        ],
        "summary": "List Clips",
        "operationId": "list_clips_api_v1_projects__project_id__clips_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClipOut"
                  },
                  "title": "Response List Clips Api V1 Projects  Project Id  Clips Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "clips"
        ],
        "summary": "Create Clip",
        "operationId": "create_clip_api_v1_projects__project_id__clips_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClipCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClipOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clips/{clip_id}": {
      "get": {
        "tags": [
          "clips"
        ],
        "summary": "Get Clip Detail",
        "operationId": "get_clip_detail_api_v1_clips__clip_id__get",
        "parameters": [
          {
            "name": "clip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Clip Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClipOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "clips"
        ],
        "summary": "Update Clip",
        "operationId": "update_clip_api_v1_clips__clip_id__patch",
        "parameters": [
          {
            "name": "clip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Clip Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClipUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClipOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "clips"
        ],
        "summary": "Delete Clip",
        "operationId": "delete_clip_api_v1_clips__clip_id__delete",
        "parameters": [
          {
            "name": "clip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Clip Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clips/{clip_id}/duplicate": {
      "post": {
        "tags": [
          "clips"
        ],
        "summary": "Duplicate Clip",
        "operationId": "duplicate_clip_api_v1_clips__clip_id__duplicate_post",
        "parameters": [
          {
            "name": "clip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Clip Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClipOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clips/{clip_id}/render": {
      "post": {
        "tags": [
          "clips"
        ],
        "summary": "Render Clip",
        "operationId": "render_clip_api_v1_clips__clip_id__render_post",
        "parameters": [
          {
            "name": "clip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Clip Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clips/{clip_id}/exports": {
      "post": {
        "tags": [
          "clips"
        ],
        "summary": "Create Export",
        "description": "Render an MP4 export of the clip with its current settings (same as /render).",
        "operationId": "create_export_api_v1_clips__clip_id__exports_post",
        "parameters": [
          {
            "name": "clip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Clip Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clips/{clip_id}/download": {
      "get": {
        "tags": [
          "clips"
        ],
        "summary": "Download Clip",
        "description": "Short-lived signed URL to the clip's most recent rendered MP4.",
        "operationId": "download_clip_api_v1_clips__clip_id__download_get",
        "parameters": [
          {
            "name": "clip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Clip Id"
            }
          },
          {
            "name": "inline",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Inline"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignedUrlOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/clips/{clip_id}/captions": {
      "post": {
        "tags": [
          "clips"
        ],
        "summary": "Generate Captions",
        "operationId": "generate_captions_api_v1_clips__clip_id__captions_post",
        "parameters": [
          {
            "name": "clip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Clip Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "503": {
            "description": "Captions are not available yet"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exports": {
      "get": {
        "tags": [
          "clips"
        ],
        "summary": "List Exports",
        "operationId": "list_exports_api_v1_exports_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_MediaOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exports/{export_id}": {
      "get": {
        "tags": [
          "clips"
        ],
        "summary": "Get Export",
        "operationId": "get_export_api_v1_exports__export_id__get",
        "parameters": [
          {
            "name": "export_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Export Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignedUrlOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClipCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "start_seconds": {
            "type": "number",
            "minimum": 0.0,
            "title": "Start Seconds"
          },
          "end_seconds": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "End Seconds"
          },
          "render_settings": {
            "$ref": "#/components/schemas/RenderSettings"
          }
        },
        "type": "object",
        "required": [
          "title",
          "start_seconds",
          "end_seconds"
        ],
        "title": "ClipCreate"
      },
      "ClipOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "start_seconds": {
            "type": "number",
            "title": "Start Seconds"
          },
          "end_seconds": {
            "type": "number",
            "title": "End Seconds"
          },
          "duration_seconds": {
            "type": "number",
            "title": "Duration Seconds"
          },
          "selection_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selection Reason"
          },
          "engagement_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engagement Score"
          },
          "origin": {
            "type": "string",
            "title": "Origin"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "transcript_excerpt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transcript Excerpt"
          },
          "render_settings": {
            "additionalProperties": true,
            "type": "object",
            "title": "Render Settings"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "render_is_current": {
            "type": "boolean",
            "title": "Render Is Current",
            "default": false
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Url"
          },
          "latest_job": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "title",
          "start_seconds",
          "end_seconds",
          "duration_seconds",
          "selection_reason",
          "engagement_score",
          "origin",
          "status",
          "transcript_excerpt",
          "render_settings",
          "created_at",
          "updated_at"
        ],
        "title": "ClipOut"
      },
      "ClipUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "start_seconds": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Seconds"
          },
          "end_seconds": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "End Seconds"
          },
          "render_settings": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RenderSettings"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ClipUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "clip_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clip Id"
          },
          "job_type": {
            "type": "string",
            "title": "Job Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress": {
            "type": "number",
            "title": "Progress"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "attempt_count": {
            "type": "integer",
            "title": "Attempt Count"
          },
          "max_attempts": {
            "type": "integer",
            "title": "Max Attempts"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "safe_error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Safe Error Message"
          },
          "cancel_requested": {
            "type": "boolean",
            "title": "Cancel Requested"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "clip_id",
          "job_type",
          "status",
          "progress",
          "stage",
          "attempt_count",
          "max_attempts",
          "error_code",
          "safe_error_message",
          "cancel_requested",
          "created_at",
          "started_at",
          "completed_at",
          "updated_at"
        ],
        "title": "JobOut"
      },
...
      "Page_MediaOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MediaOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[MediaOut]"
      },
...
      "SignedUrlOut": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          }
        },
        "type": "object",
        "required": [
          "url",
          "expires_in"
        ],
        "title": "SignedUrlOut"
      },
...
    }
  }
}